Refactor core logic out of handler
